+++ libc/src/math/generic/log10.cpp
@@ -917,7 +917,6 @@ double log10_accurate(int e_x, int index, double m_x) {
 
   // Further range reductions.
   double scale = 0x1.0p+7;
-  const fputil::DyadicFloat<128> NEG_ONE(-1.0);
   for (size_t i = 0; i < R_STEPS; ++i) {
     scale *= 0x1.0p+4;
     int id = static_cast<int>(fputil::multiply_add(mx.hi, scale, 0x1.0p+4));